+++ DerailValleyDesktopGauge/Main.cs
@@ -2,15 +2,17 @@
 using System.Reflection;
 using HarmonyLib;
 using UnityModManagerNet;
-using UnityEngine;
 
 namespace DerailValleyDesktopGauge;
 
+#if DEBUG
+[EnableReloading]
+#endif
 public static class Main
 {
     public static UnityModManager.ModEntry ModEntry;
-    public static Settings Settings;
-    private static OpenSimGaugeManager _openSimGaugeManager;
+    public static Settings settings;
+    public static OpenSimGaugeManager openSimGaugeManager;
 
     private static bool Load(UnityModManager.ModEntry modEntry)
     {
@@ -19,16 +21,15 @@ public static class Main
         Harmony? harmony = null;
         try
         {
-            Settings = Settings.Load<Settings>(modEntry);
-
-            // modEntry.OnGUI = OnGUI;
-            // modEntry.OnSaveGUI = OnSaveGUI;
+            settings = Settings.Load<Settings>(modEntry);
+            modEntry.OnGUI = OnGUI;
+            modEntry.OnSaveGUI = OnSaveGUI;
 
             harmony = new Harmony(modEntry.Info.Id);
             harmony.PatchAll(Assembly.GetExecutingAssembly());
 
-            _openSimGaugeManager = new OpenSimGaugeManager();
-            _openSimGaugeManager.Start();
+            openSimGaugeManager = new OpenSimGaugeManager();
+            openSimGaugeManager.Start();
 
             ModEntry.Logger.Log("DerailValleyDesktopGauge started");
         }
@@ -43,23 +44,19 @@ public static class Main
         return true;
     }
 
-    // private static void OnGUI(UnityModManager.ModEntry modEntry)
-    // {
-    //     GUILayout.Label("Mod Settings", UnityEngine.GUI.skin.label);
+    private static void OnGUI(UnityModManager.ModEntry modEntry)
+    {
+        settings.Draw(modEntry);
+    }
 
-    //     Settings.Port = int.Parse(
-    //         GUILayout.TextField(Settings.Port.ToString())
-    //     );
-    // }
-
-    // private static void OnSaveGUI(UnityModManager.ModEntry modEntry)
-    // {
-    //     Settings.Save(modEntry);
-    // }
+    private static void OnSaveGUI(UnityModManager.ModEntry modEntry)
+    {
+        settings.Save(modEntry);
+    }
 
     private static bool Unload(UnityModManager.ModEntry entry)
     {
-        _openSimGaugeManager?.Stop();
+        openSimGaugeManager?.Stop();
         ModEntry.Logger.Log("DerailValleyDesktopGauge stopped");
         return true;
     }