+++ DerailValleyDesktopGauge/OpenSimGaugeManager.cs
@@ -1,100 +1,53 @@
 using System;
-using System.Linq;
 using System.IO;
-using System.IO.Compression;
 using System.Diagnostics;
+using UnityModManagerNet;
+using System.Linq;
 
 namespace DerailValleyDesktopGauge;
 
-public class OpenSimGaugeManager {
-    
+public class OpenSimGaugeManager
+{
+    private static UnityModManager.ModEntry.ModLogger Logger => Main.ModEntry.Logger;
     private static Process? _externalProc;
-    private static string DepsPath = Path.Combine(Main.ModEntry.Path, "Dependencies");
-    private static string ZipExtractPath = Path.Combine(Main.ModEntry.Path, "OpenSimGauge");
+    public static string ExeName = "OpenSimGauge";
 
-    private void Extract()
+    public bool GetIsRunning()
     {
-        Main.ModEntry.Logger.Log("Extracting...");
+        if (_externalProc != null && !_externalProc.HasExited)
+            return true;
 
-        string zipPath = GetZipPath();
+        var detectedProcess = Process.GetProcessesByName(ExeName).FirstOrDefault();
 
-        if (string.IsNullOrEmpty(zipPath))
-            throw new Exception($"No zip found in {DepsPath}");
-
-        Main.ModEntry.Logger.Log($"Extract {zipPath} => {ZipExtractPath}");
-
-        ZipFile.ExtractToDirectory(zipPath, ZipExtractPath);
-        
-        Main.ModEntry.Logger.Log("Extract done");
-    }
-
-
-    public static void CopyDirectory(string sourceDir, string destDir)
-    {
-        Main.ModEntry.Logger.Log($"Copy {sourceDir} => {destDir}");
-
-        Directory.CreateDirectory(destDir);
-
-        foreach (var file in Directory.GetFiles(sourceDir))
-        {
-            var fileName = Path.GetFileName(file);
-            var destPath = Path.Combine(destDir, fileName);
-
-            File.Copy(file, destPath, overwrite: true);
-        }
-
-        foreach (var dir in Directory.GetDirectories(sourceDir))
-        {
-            var dirName = Path.GetFileName(dir);
-            var destPath = Path.Combine(destDir, dirName);
-
-            CopyDirectory(dir, destPath);
-        }
-    }
-
-    private void ExtractAndEnhance()
-    {
-        Extract();
-
-        Main.ModEntry.Logger.Log("Enhancing...");
-
-        var enhanceDirPath = Path.Combine(DepsPath, "enhance");
-
-        CopyDirectory(enhanceDirPath, ZipExtractPath);
-        
-        Main.ModEntry.Logger.Log("Enhance done");
-    }
-
-    private string GetZipPath()
-    {
-        Main.ModEntry.Logger.Log($"Searching for ZIP in: {DepsPath}");
-
-        string? firstZip = Directory
-            .EnumerateFiles(DepsPath, "*.zip", SearchOption.TopDirectoryOnly)
-            .FirstOrDefault();
-        return firstZip;
-    }
-
-    private bool IsExtracted()
-    {
-        return Directory.Exists(ZipExtractPath);
+        return detectedProcess != null;
     }
 
     public void Start()
     {
-        Main.ModEntry.Logger.Log("Starting OpenSimGauge...");
+        Logger.Log("Starting OpenSimGauge...");
 
-        if (!IsExtracted())
+        var isRunning = GetIsRunning();
+
+        Logger.Log($"Is already running: {isRunning}");
+
+        if (isRunning)
         {
-            ExtractAndEnhance();
+            Logger.Log("Cannot start: already running");
+            return;
         }
 
-        if (!IsExtracted())
+        if (!EnhancerExtractor.IsExtracted())
+            EnhancerExtractor.ExtractAndEnhance();
+
+        if (!EnhancerExtractor.IsExtracted())
             throw new Exception("Failed to extract");
+
+        if (EnhancerExtractor.NeedsMoreEnhancement())
+            EnhancerExtractor.EnhanceSafely();
 
         var psi = new ProcessStartInfo
         {
-            FileName = Path.Combine(Main.ModEntry.Path, "OpenSimGauge/OpenSimGauge.exe"),
+            FileName = Path.Combine(Main.ModEntry.Path, $"OpenSimGauge/{ExeName}.exe"),
             UseShellExecute = false,
             CreateNoWindow = false,
         };
@@ -104,33 +57,41 @@ public class OpenSimGaugeManager {
 
     public void Stop()
     {
-        Main.ModEntry.Logger.Log("Stopping OpenSimGauge...");
+        Logger.Log("Stopping OpenSimGauge...");
 
         try
         {
             if (_externalProc != null && !_externalProc.HasExited)
-            {        
-                Main.ModEntry.Logger.Log("Closing main window...");
+            {
+                Logger.Log("Closing main window...");
 
                 _externalProc.CloseMainWindow();
 
                 if (!_externalProc.WaitForExit(3000))
                 {
-                    Main.ModEntry.Logger.Log("Failed to close main window - killing process...");
+                    Logger.Log("Failed to close main window - killing process...");
                     _externalProc.Kill();
-                    Main.ModEntry.Logger.Log("Done");
+                    Logger.Log("Done");
                 }
-                
-                Main.ModEntry.Logger.Log("Closed");
+
+                Logger.Log("Closed");
             }
             else
             {
-                Main.ModEntry.Logger.Log("Already closed");
+                Logger.Log("Already closed");
             }
         }
         catch (Exception ex)
         {
-            Main.ModEntry.Logger.LogException("Failed to close external process:", ex);
+            Logger.LogException("Failed to close external process:", ex);
         }
+    }
+
+    public void Restart()
+    {
+        Logger.Log("Restarting OpenSimGauge...");
+
+        Stop();
+        Start();
     }
 }